Adjust effect damage and enchantment of the mace of variability
The mace of variability casts chain of chaos with spell-power ranging
from 75 to 149 with an average of 112. Ignoring any damage from the
chaos beam effect, the current arc damage formula of 3d(5 + pow / 2)
gives the mace's spell damage die of 3d61 per arc on average. This is
very high, especially since the monster gets zapped 4 times on average
(albeit with spell-power that decreases by 14 on average with each arc).

To balance the damage from the mace, reduce the damage formula of chain
of chaos to 3d(5 + pow / 6). Now an arc from the spell will do 3d23
damage on average, which is still significant damage even when cast only
1 in 5 melee hits. And of course, in addition to damage, each arc
applies random effect from the chaos beam.

This commit also lowers the enchantment of the mace to +7 from a
previous commit's value of +10. A +7 great mace will be quite strong
enough with 1 in 5 attacks casting chain of chaos in addition to the
normal chaos ego. Also 7 is a lucky number, which Xom appreciates, and a
value lower than the typical +9 max is a **secret reference** to the
original unrand behaviour.