Adjust the effects of chaos beams
The chaos beam type is only used for the chain of chaos spell cast by
the mace of variability and for the impact (non-cloud) effect of
apocalypse crab breath. Currently chaos beams select from a set of beams
analogous to effects used by the chaos brand, but using mostly flat
selection weights. The flat weights effectively give chaos beams much
higher weighting relative to the chaos brand for strong effects like
banishment, polymorph, hexing or buffing the monster, or miscast
effects. This difference isn't very apparent for chaos breath, since the
beam is only applied once for impact damage, but for the mace of
variability, a chaos arc can hit a monster multiple times in a turn.

This commit re-weights the chaos beam effect table to be closer in
behaviour to the chaos brand. Beams directly analogous to other brands
like flaming, freezing, electrocution, etc now have comparatively higher
weight, and rarer effects that only happen when the chaos brand
reselects itself have lower weight.

Note that this commit has no effect on the behaviour of the chaos brand,
monster af_chaos, or chaos clouds; it's only relevant for the mace of
variability and impact (non-cloud) effects from chaos breath. It would
be ideal to somehow merge the selection of these beam effects with the
effect selection used for SPWPN_CHAOS, especially since the latter is
already used for all chaos melee and launcher egos as well as chaos
clouds.