Remove unused monster casting for chain of chaos
This spell used to be castable on a demonspawn monster, but that monster
no longer exists. Chain of chaos isn't very good as a monster spell
since it applies multiple chaos effects to the player per turn and at
range. Chaos breath is ranged but only applies on effect per turn due to
its use of clouds, so this is at least a better style of spell or breath
to give monsters.

This commit removes code used to let monsters cast chain of chaos
directly. Monsters can still use the chain of chaos effect from wielding
the mace of variability, since that triggers from a melee effect and
doesn't need the usual monster spell setup functions.