Revise the chance for pulling players deeper into the Abyss
The new formula removes the XL15 cutoff for the chance to pull the
player deeper into the Abyss upon shift/teleport. The chance formula is
now:

  (XL + 4 - Depth)^2 / (27^2 * (Depth + 5))

Some of the new chances:
* Characters at XL 1 have about a 0.1% chance of getting pulled from
  A:1.
* Characters at XL 13 have chances of 5.9%/4.4%/3.3%/2.6% for getting
  pulled from A:1/A:2/A:3/A:4.
* Characters at XL 27 have chances of 20.6%/16.5%/13.4%/11.1% over the
  same levels.

It's relatively easy to escape from the abyss at XL13, and a small
chance of getting pulled a level deeper where exits are more common
won't have much impact on players at these XLs.

This commit also updates comments and spacing and properly declares a
function as static. Closes #757.