Give non-menu abilities default flags, document, and reformat
Give Wu Jian's Lunge and Whirlwind abilities default flags so people
don't think that these actually show up in the menu. It might be nice if
the ability enums/entries didn't have to exist just because of action
counting, but that would take some reworking of the action counting
system. Closes #775.