Don't dissipate clouds when testing door sealing for tension (11506)
As |amethyst suggested, this is happening because clouds will be
instantly deleted on terrain changes that leave them out of los (as of
7e89378e288a). The door sealing code tests various door closing options
to see which maximizes tension, and these would delete the clouds even
if none of them for used, because their los effect is still calculated.
This solution simply backs up a copy of env.cloud, since in principle
the cloud changes might impact tension.

(cherry picked from commit acb6733a4425a3cb941e387e420bd0349c60805c)