Remove code for using decks when worshiping gods other than Nemelex
Back when decks could appear as random items and were retained upon
abandonment of Nemelex, cards could potentially be used by players
worshiping other gods. Hence there is code in place to prevent
generating cloud types that specific gods disapprove of and to track use
of Degeneration under Zin.

This commit removes this code, since decks no longer generate randomly
and are destroyed upon abandonment of Nemelex. It could have a
TAG_MAJOR, but the gameplay ramifications of this change are very small.