A stat drain themed ghost vault for Dungeon and Orc (mikee)
Players who don't watch out for those pesky sources of stat drain might
end up haunting the transporter vault gammafunk_ghost_stat_death.
Dex-draining Quasits are the only threat accompanying the ghost early
on, with the quasits being supplemented by int-draining neqoxecs by D:9.
There's the usual pile of progressive loot, and the ghost generates over
a ring of dexterity or intelligence that has progressively higher
chances for being good_item or randart.

Based on an initial suggestion from mikee for a ghost vault with a bunch
of quasits and a ring of DEX.