Add a scarf of stasis (#38)
This commit adds a new scarf: the scarf of stasis. This aims to add
interesting decisions to the game through offering players access to
stasis and partially mitigating the biggest MR threat in the game,
paralysis, at the cost of their most powerful consumables. It takes 5
turns to remove a scarf, which is long enough to be unviable while in
combat but short enough to be removable in situations where ambrosia
would be used, for example. Additionally, since scarves are rare, I
don't think having this step on formicids' toes is a particularly big
issue.

Here goes...