Enforce minimum buffer sizes for domain sockets
On linux these have reasonable defaults, but on OS X (and probably other
unixes) the receive buffer default for unix datagram sockets is
unreasonably small for crawl's purposes, 4096. This mins the receive
buffer at linux's default (which is enough that a crawl map can transfer
without errors). For sanity's sake I added a min for the send buffer,
but I think this value is much more standard.