Fix polymorph/frenzy bug
This issue is due to frenzy setting the attitude and storing the old
attitude in the props map and it never gets set back when polymorphed as
it simply clears the enchantment map. The old attitude usually gets
restored when remove enchantment is called after the duration times out
but since that isn't called it keeps the indifferent attitude.

This is mostly a band-aid fix to just fix this one instance because
other enchantments don't seem to have similar issues. A more long term
fix would be adding a function that calls remove enchantment function
for each enchantment. I wasn't sure the best way to do this so I have
submitted this simple fix.