Remove hair/beard layer from Ba, VS, Gr
Displaying hair on these tiles was unintentional.

Ba for example is drawn to have 4 eyes, but the default hair covers up
the top two. That, and considering the fact that frogs aren't hairy IRL
makes it pretty clear that displaying hair was an accident, not a
deliberate artistic decision.

This commit removes Gnolls from this commit, since it does not make
sense for gnolls, a hyena-human hybrid, to have no hair.