A mace-themed ghost transporter vault for Dungeon and Orc
Various ogre and giant monsters wait inside a transporter area to remind
you to respect things that can hit with a giant stick, lest you end up
like the ghost in this vault. Places ogres initially, progressing to
two-headed ogres, with a chance to place an ettin, frost giant, or fire
giant in the latest depths. One monster of the most dangerous type will
have a fancy good_item or randart mace that is either a dire flail or
great mace. If this monster is a fire or frost giant, its mace will be
flaming or freezing, respectively.