Have transmutations last forever at high spell power (PR #23)
It is rather tedious to have to recast transmutations constantly, so this commit makes it so that at high power, transmutations last forever (currently 1 bar of power per spell level; this may need adjusting for balance reasons if it's too easy to attain this). This applies to the five "form" transmutations, and not appendage or blade hands.