Quick fix for webtiles menu bug
The problem seems to be that this global variable was getting out of
date at some point, and so referenced a div that is no longer in the
DOM.  I haven't tracked down why but at least on old UI code this sort
of thing would happen when resizing, since the DOM would get completely
rebuilt. In any case, in my testing this fixes the bug.