Updates to gammafunk_ghost_fury_of_okawaru
Place Okawaru gift inspired weapon and armour items for the first two
loot pieces in this vault instead of the usual superb or star items. The
weapon item uses the good weapon base types, with lower weight on the
rarer types, and is always good_item. The armour is more general armour
but can either be good_item or randart. The weapon item always places,
but the armour item only replaces a superb/star item 1/3 of the time
initially, with that chance increasing to 100% as depth increases.

Also place a bad weapon or armour item (negatively enchanted, non-ego,
cursed) under the ghost, since we all know this is what Oka gifts
usually look like.  This bad item is always pre-identified.