Fix crash when scrolling+switching in webtiles menus
A webtiles client may send a scroll event before it is informed of
updates to menu content. If the number of menu items is then changed,
such scroll events may reference an out-of-bounds item index by the time
they arrive at the server.

Note that any such scroll events are 'stale' in the sense that they
refer to scrolling to an item/offset that is no longer shown, so there
is no way for servers to handle these events. A more robust solution
would involve webtiles clients acknowledging all updates to menu
content, and the server side using the ordering of acknowledgement and
scroll update messages to detect stale scrolls.