Don't overwrite god with GOD_NAMELESS on poly to a priest (11536)
This was creating duplicate monsters under the same mid in the following
way for beogh worshippers: 1. an orc follower gets turned into a
shapeshifter via chaos effects (e.g. an apocalypse crab). 2. the monster
shapeshifts into a priest monster, such as a meliai, setting the
monster's god to GOD_NAMELESS. 3. The player exits the level, and
recalls allies on a new level. A copy of the follower is generated at
this point. 4. They reenter the level of step 2. Because the monster
copy on that level no longer worships Beogh, is_divine_companion returns
false, and the follower copy is not appropriately cleaned up when
re-entering that level (see `tag_read_level_monsters`). Now there are
two. (Probably other variants of this procedure could work.)

This state doesn't crash, but the situation can eventually lead to the
infamous "mid cache bogosity" message in various ways, such as if one of
the copies dies while both are on the level, removing it from the mid
cache.

The fix might give priest monsters some weird gods at times, but it
doesn't mess up orc followers.