Adjust some and add 1 new ghost vault
This is a squashed version of three commits.

 52fbc5fa17 Rename a ghost vault

Rename ploomutoo_gammafunk_potion_laboratory to
ploomutoo_gammafunk_ghost_potion_laboratory so that it has 'ghost' in
the name of the vault like the other ghost vaults do.

 7ef8c7bd50 Resize a ghost vault

Currently, any monsters and loot at the back of the vault
gammafunk_ghost_cemetery are one tile out of LOS even when viewed from
outside. This commit decreases the size of the vault so that all floor
tiles can be seen.

 038298e604 A kobold themed ghost vault for Dungeon

In gammafunk_ghost_nasty_kobolds, the kobolds pull out all the nasty
tricks: blowguns, wands, daggers with distortion or other nasty brands,
and demon summoning. Places a fancy kobold with a kobold-appropriate 1h
weapon that's branded and good_item or randart. You'll also find at
least one of the aforementioned distortion weapon kobolds. Additionally
there's at least one kobold with either a blowgun with poison or curare
needles or one with a monster-usable wand. Numbers of these increase
with depth, with big kobolds replacing ordinary kobolds and with kobold
demonologists gradually being introduced. Who knows, maybe the lone
ghost you'll find inside is the result of someone failing to appreciate
just how tricky and nasty dungeon kobolds can be.