Preserve insanity on polymorph (11504)
Because insanity involves an attitude change, polymorph was removing the
enchantment and not returning the attitude to normal. This changes
leaves the monster insane after a polymorph, and the enchantment goes
away as normal, returning the attitude to its correct state at that
point.

(cherry picked from commit 337e1da0210d9ee37bb51596f28be351124fbf21)