Save compat for the Beogh/priest polymorph bug
See e6d7efa92cb0cd for the bug itself.

This cleans up duplicate mid monsters for various cases that match the
conditions of this bug. Possibly this should have a minor version tag
bump, but as of right now that will mess up the ghost permastores and I
don't want to do that this close to a release.