Avoid "mid cache bogosity" for unhandled divine clone bugs
Companion clone bugs lead to "mid cache bogosity" under the following
circumstances:

1. there is a level saved with a copy of the companion that gets false
for is_divine_companion() (e.g. it lost its god, as in e6d7efa92cb0).
2. that level also has a properly saved copy of the companion; they
both share the same mid.
3. the mindex of the corrupted copy is lower than the correct copy.

If all three conditions are met, the elsewhere companion code would
clear the mid cache for the corrupted copy, but the corrupted copy would
still be alive and kicking in the monster list. This would lead to a
crash when entering the level, resulting in an unplayable save.

This commit avoids this situation by fixing up the mid cache after
deleting the elsewhere companion. This fix will still result in errors
under various circumstances, but hopefully not unplayable saves. An
alternative might be to always delete both copies, but doing so will
hide whatever the underlying bug is that is leading to a clone so IMO
it's better to print lots of errors when these things happen, and delete
only corrupted copies where we know what the bug was.

As a side note, if there are multiple copies but they satisfy
is_divine_companion, they'll still be silently deleted -- this is the
status quo from before this commit. If condition 3 above isn't met,
there will also be no error, on load at least.