Adjust Venom Mage
This commit:
  - Removes Mephitic Cloud from the Young Poisoner's Handbook and
    replaces it with Ignite Poison;
  - Buffs the power scaling of Ignite Poison;
  - Buffs the cloud duration of Poisonous Vapours;
  - Adds a new spell for Venom Mages:
    level 4 poison/summoning Summon Scorpions, which summons one or more
    scorpions with a summon cap of 1 band.
  - Ensures that Sting and Venom Bolt always poison.
  - Allows Poisonous Vapours to be cast and create a cloud anywhere.

These changes are to make Venom Mage a stronger background that relies
less on Mephitic Cloud, which is boring, also applies to Wizard, and
ends up encouraging weapon-based combat; and relies more on actual
poison-related effects to kill enemies. It also reduces reliance on
melee combat and finding other spells by adding the Ignite Poison and
Summon Scorpion spells, which allow the player to, in a very crude way,
deal with poison resistant and immune creatures.

This also will hopefully make VM feel more unique and entice more
people to play what was a very tedious background.

This commit merges #44.