Fix webtiles popups obscuring direction chooser (stoneychips, #11525)
This issue was visible when evoking items from their item description,
opened via the inventory. It wasn't a problem on console or local tiles
because the direction chooser code runs its own render loop. The fix
here is analagous to the UI_VIEW_MAP UI state used for viewing the map.