Work around direction-chooser console rendering bug
This rendering bug occurs when viewing an item's description via the
inventory, and then evoking it: the stats sidebar of the main dungeon
view UI is not re-rendered, and instead shows a portion of the item
description text.

I don't feel comfortable calling this a fix, because it's rather
ham-fisted. It's positioned after the clear_messages() call to avoid
causing flicker in the messages panel.