Improve scroll shade appearance (pedant, #11520)
The items obscured by scroll shade gradients were too easily
misinterpreted as being greyed-out. This commit fixes that by using a
much narrower gradient band. This effect was compounded on webtiles
where the gradient was darker: the webtiles CSS has been changed to
match the local tiles colours.

The gradient band was originally taller to indicate the presence of
additional content beyond the edge of the scroller viewport, but with
the addition of scrollbars to (web)tiles, this is less important.