Buff the Vampire's tooth
This is a squashed version of two commits. We believe the Vampire's
tooth was so utterly useless that this improvement does not require a
vote to be pulled.

 b41cb9a525 Buff Vampire's tooth (gammafunk)

The gimmick (100% of damage comes as vampiric heal) only works on a
weapon that's able to do some damage. This patch changes it from a +4
dagger to a +12 quickblade. The quick attack and slaying give it an
average damage less than a +4 electric brand quick blade, but enough
healing to be similar to a spellpower 100 vampiric draining cast against
targets that aren't too heavily armoured.

 dfd0697590 Update the Vampire's Tooth description