Fix segfault after viewing help from menu (josh, #11553)
These extra lines were holdovers from the previous menu implementations,
which needed to explicitly re-render after showing the help screen and
used reentrancy to do so. This caused an additional UI layer to be
pushed for every time the help was opened; these were not popped when
end() was called, because the UI framework does not handle exceptions
(yet?).