Remove rC- and rF- from Maxwell's Thermic Engine
This weapon has damage roughly similar to a +9 Freezing or Flaming
double sword, considering how much its enchantment fluctuates during a
fight. Assuming you can sustain its highest enchantment and the target
has neither rC nor rF, it's the same damage as +14 speed double sword.
This is outstanding, but not to the point where the item needs something
like negative resistances or other drawbacks to balance it. The negative
resistances also aren't an interesting trade-off on a swappable weapon.