Simplify Beogh conversion rules
The previous commit mentions that these lead to a minigame of 'don't
attack this orc until you're certain its met the simulationist criteria
for a conversion attempt'. This commit makes orc conversion attempts
happen on sight, like Gozag and Jiyva (even if the orc is asleep,
confused, paralyzed, can't see you, or has not yet selected you as their
foe), removing the minigame.