Improve installation using MSYS2 MINGW64
This is a squashed version of five commits. Most excitingly, this commit
fixes the sdl2 submodule to finally allow the compilation of the tiles
version of Crawl (and now X-Crawl) on Windows. After pulling this
commit, it is necessary to run `git submodule update`. Please do
so before attempting to compile tiles.

 4a1e0138fc Improve MSYS2 installation instructions

Previously, this made the reader skip around to various other parts of
this document and other web pages; this commit tries to lay everything
out in steps.

 8522bf8125 Update submodules to improve windows compilation

This has another stab at the xinput issue, to deal with further changes
(apparently) in MSYS2. [This finally completely fixes Windows tiles
compilation.]

 05cf915bf0 checkwhite INSTALL.txt
 8a29c8c6af Further tweaks to MSYS2 instructions in INSTALL.txt
 ec0ffd4cd7 clarify what is recommended for windows builds (gammafunk)