Pad textures to powers of two (aidanh)
This ensures that the texture height is always a power of two, leaving
the width still fixed at 1024. This may address issues on some older
hardware/drivers (and potentially VMs) where textures don't load
properly.

If *square* power-of-two textures are needed that will be a bit more
effort, since the 1024 width seems to be a bit hard-coded into things.