Allow replay_messages from main menu with ctrl-p
This addresses a long-standing complaint of mine: it's a pain to see
debug messages that happen before game start. With the new UI system,
this is the kind of thing that's trivial to add. For regular players,
this will mainly be in case some error messages show up, so it's not
documented in the menu interface.

This commit also tries to more aggressively avoid more prompts during
game startup; for those of us with 50 save games running debug builds,
this makes startup a bit smoother -- what information you could get out
of this can now be seen in ctrl-p.