Notify user if there were errors during initialization
This will help with things like the white tile issue: now there should
be opengl errors on startup and this will tell the user about them, so
that they user will have some better idea of who or what to blame.

Possibly it would be better to force a popup with the errors, but this
is a bit more work and the replay_messages ui would need some tweaking
for that use case (I first implemented it this way, but it was too
cryptic), so I'm saving that for a later version.