Optimize memorize menu a bit
I was getting noticeable lag on opening the spell menu, and profiling
revealed that sorting was having a huge impact, because the sort
function collected all sorts of spell information each time it was
called. The main change of this commit is to precalculate spell
information for each spell in the menu, since they won't change while
memorizing.  This speeds things up quite a bit for me (bringing things
down from ~200ms to ~50ms for a big memorize list). The biggest offender
was repeated calls of failure_rate_to_int(raw_spell_fell(...)). This may
be exaggerated in debug mode since part of the hit was that each call of
that does a dprf, but that wasn't the whole story in the profiling data,
so I expect this will help in non-debug builds as well.

I also hit a few minor slower points including the status line
construction from 985ef60fe1445, and did a bunch of minor modernization
of the c++ style in the memorize menu code in general.