Set expand_v on memorize menu
This is a quick change to keep the status lines in the same place when
searching, rather than dynamically resizing the menu. Possibly the max
sizes could use more tweaking (it might make sense to set the max size
based on the initial spell list, since it can never grow). Also, this
doesn't work on webtiles; resizing on search looked the worst on console
so that's probably ok.