Revise Morg (mikee)
Morg was a unremarkable flavour unrand, serving as Boris' knife when he
was alive. This patch attempts to keep the flavour but also make Morg
interesting. This is done by making Morg's pain use 1.5 * Necromancy for
the damage calculation, giving it Harm and *Drain. The idea is that this
knife should be extra evil. The 1.5 * Necromancy alone gives Morg a lot
of new damage, increasing the plus might be too much.

Some discussion suggests that Harm might not be the best (at least, it's
frowned upon by some players), in particular that the kind of character
that has a lot of necro won't want a Harm weapon because they lack
defenses. However, I think Harm makes Morg an interesting choice: is the
danger and lack of enhancer access worth the extra damage and XP
savings? For a Kiku necromancer, probably not, as the damage turns out
to be only slightly more than a +9 quick blade of pain. For a
necromancer without Kiku's 6* gift it seems like it depends a lot on the
character.