Revise the Spriggan's Knife (mikee)
Another flavour unrand. The knife now always stabs at Tier 1;
specifically any stab is converted to a sleep stab, this is flavoured as
cutting into the realm of dreams. Remove the Dex+4 and MR+ to de
stat-stick it a little. The Dex+4 in addition to the EV and Stealth
plusses was overkill, and since the knife now always stabs at tier 1 the
extra Dex seemed like too much.

A separate idea discussed for the knife was to make it a Hex enhancer.
There's probably room for a Hex enhancer unrand dagger also.