Ensure that clearing a menu triggers a layout
In local tiles, a pack_buffers call is necessary to relayout a menu, but
simply calling deleteAll(items) doesn't set m_needs_layout (which
triggers the sequence of events that leads to pack_buffers). Console
works differently for reasons I haven't examined closely (this bug
didn't affect it).  This commit calls _queue_allocation on clear() in
order to have a way of clearing the menu and triggering a layout. This
may pop up in other menus since the deleteAll paradigm is used in a
bunch of places.

This was manifesting for the memorize menu for the case where the menu
was cleared during update_entries but no items were added to it.