Revise the Singing Sword
Make the sonic damage occur on-hit, instead of as a tension gated
effect. The sword becomes the screaming sword when there are any visible
monsters, and each turn has a chance of making some noise (using the low
tension message). On a hit, it has a two in three chance of making a
noise. The noise has a one in three chance of being a scream and doing
double damage, and otherwise gives a high tension message and does
regular damage.

The damage formula has been simplified, it no longer takes into account
monster mhp or player XL. Instead, a normal noise is 1d2 * 1d2 * 1d3 +
1d3 + 1, and a scream doubles it. This gives a damage function with mean
10, deviation 5.63, and max 32. The distribution is not normal, 90% of
rolls are less than or equal to 16. This is a significant buff for the
sonic damage, but still a lot less than Ozocubo's Refridgeration, and
very noisy (25 or 35 each hit!).