Update ?? screen and hints mode for spell library (11561)
This is just a quick pass, and probably needs some more work; in
particular the hints mode text for spellcasting is now kind of buried in
the spell library help screen and needs to be brought out a bit more.

It doesn't seem like ?? really got updated for floor actions either, so
I took that into account on my rearrangement.