Adjust Jorgrun's spells and move him earlier
Shatter is not a good monster spell, since it can often one-shot
Gargoyle players or players who happen to move into LOS in statue form.
There are enough LOS-wide unavoidable damage monster spells, and we
don't need another one just because it's based on a player spell or
because it can be resisted with flight. This commit removes Jorgrun's
Shatter spell, replacing it with Iron Shot. Iron shot is not the most
interesting spell for a unique to have, but it does give him an
always-usable source of ranged damage. I have a local branch testing out
monster Borg's Vile Clutch, but this is probably a silly idea. Monsters
with ranged constriction could be fun, so I'll test it and possibly try
replacing his Iron Shot with that spell in 0.23. Should this unlikely
event happen, any blame will deftly be redirected to mikee.

This commit also adjusts Jorgrun's placement to have him show up earlier
and on fewer levels. Previously he showed up on 30 levels, second only
to Donald. His new placement range is:

D:12-15, Orc, Elf:1-2, Snake:1-3, Shoals:1-3, Spider:1-3

His Petrify spell is more effective when players are less likely to have
high MR, so using these earlier levels will help his Petrify/LRD combo
be more relevant. Swamp is omitted since his Dig and LRD are not
effective there.

We are pulling this commit to X-Crawl in its entirety, because this is
what X-Crawl voters ended up voting for anyway. It is nice that mainline
Crawl has also recognised that this is the way to go, though.