Fix messaging with cloud of scarf and qazlal (11500)
This incorporates elements of both #799 and #804. Though apparently both
MarvinPA and amalloy were against it for reasons I don't fully
understand (see note in #804), I think it is preferable to show a
message when you wear this scarf when it is completely useless. It's not
exactly a lot of message spam, since this is an uncommon item, and this
kind of information is useful to new players.

Resolves #799, resolves #804