Fix messaging with scarf of cloud immunity and qazlal (11500)
This incorporates elements of both #799 and #804. Though apparently both
MarvinPA and amalloy were against it for reasons I don't fully
understand (see note in #804), I think it is preferable to show a
message when you wear this scarf when it is completely useless. It's not
exactly a lot of message spam, since this is an uncommon item, and this
kind of information is useful to new players.

This commit message has been amended because the original title referred
to a "cloud of scarf". Whether this is referring to a secret new cloud
that suffocates the player, or rather one that wraps the player up to
make them warm, I'm not sure, but one thing is certain: it isn't what
the commit is talking about.