Fix the spell targeter for Gell's Gravitas
Previously, the formula used to calculate the targeted area was
completely different to that which the spell itself used, resulting in
the targeter blatantly lying to the player about the tiles which could
and could not be affected.

The current behaviour of Gell's Gravitas is entirely predictable based
on spell power, with no randomness involved, therefore the new targeter
is entirely in the darker yellow colour to indicate the certainty. The
power breakpoints are:
>0 power: at radius 1 from the centre, the monster is pulled 1 tile;
>0 power: at radius 2, the monster is pulled 1 tile;
>60 power: at radius 2, the monster is pulled 2 tiles; and
>80 power: at radius 3, the monster is pulled 1 tile.

The previous spell targeter suggested that, at low power levels,
monsters at radius 2 could never be affected, and that at exceptionally
high power levels, targets at radius 4 could be affected, both of which
are incorrect. Hence, this commit changes the targeter to reflect the
fact that monsters at radius 2 are always affected and monsters at
radius 3 are always affected when, and only when, above 80 power.

An argument could be made for having the targeted tile (the centre of
the "gravitational reorientation") not appear coloured, given that this
tile is not affected, however this would require significantly more
work.

This commit removes the gravitas_range function, which was previously
only used for making the targeter, because it is no longer required.
Additionally, this commit removes the player's tile being highlighted,
given the player is never affected.