Try again to merge centaur and naga bardings
 183a67f Revert "Revert "Merge centaur and naga bardings""

This reverts commit ff05dbde86e94ea63abd27f02265f2260d0723c7.

This commit adds a new item, a barding, to replace centaur bardings and
naga bardings. These items are already stupid and already rare, so from
a gameplay standpoint it seems better for them to be merged totally. The
flavour is certainly a bit dodgy...

 c12befe Fix linesprint.des

Sprint why do you do this to me...

 83e63d7 Fix compilation in tilepick-p.cc

Creating a new tile type TILEP_BOOTS_BARDING requires a new tile, which
we don't have. Therefore, we just use the current centaur and naga
equipped barding tiles.