Improve default menu more text (fakechuck, #11563)
Showing and hiding the menu "more" text based on whether there's more
content to scroll to isn't very useful on (web)tiles; that's what the
scrollbar indicates. It's still useful on console; this commit cleans up
the default text, and shows the current scroll position as a percentage.

Currently the more text is often hijacked to display a text-based UI;
it'd be nice if these uses were separated, but that's likely somewhat
tedious.