Make Xtahua's breath apply rF-
This commit creates a new monster breath weapon, flaming breath, so that
red draconians/Tiamat don't use the same ability as Xtahua.

Now that the exclusive user of searing breath is Xtahua, we make the
searing breath ability apply fire vulnerability to the player or monster
hit by it. This makes Xtahua a stronger opponent, for two reasons: (a)
he now guards a very good set of unrandart scales that we don't want the
player getting too easily, and (b) Xtahua is who this fork is named
after, after all...