Decrease the piety cost of Qazlal's upheaval
Previously, this ability cost 3 piety. Because of the nature of Qazlal,
the player wants to spam this ability as soon as he/she gets access to
it, but that was previously not possible because you would run out of
piety too fast. This also limited the player from seeing Qaz's higher
level abilities.

Qazlal is already weak, and this commit therefore decreases the piety
cost of upheaval to 1 to make it more spammable and make the limiting
factor available MP. Happy upheaving!