Always scroll console menus down by a full page (Xenobreeder, #11507)
The issue was caused by the scroller widget automatically scrolling the
menu widget up again, so as to show as much content as possible.

This is fixed by adjusting the size requested by the menu widget,
according to how we scroll; this works because the scroller widget only
scrolls content fully into view when it is reallocated, which occurs
after the menu's size request changes.

One defect with this approach is that the size request is not adjusted
when the window (and thus the menu) is resized. This is pretty minor,
however, since console users rarely resize their terminals.