Fix webtiles not handling some scroll keys (thmsdrew, #11576)
Keypress events, such as are generated for text input, were not being
correctly delivered to scroller_handle_key(). This caused >, <, etc.
to be passed through directly to the underlying console UI, and caused
scrolling to happen on the server side. The space key was slightly
different; it seems to have been handled on both the client side
(causing an immediate scroll of one page), and then again on the server
side (causing a one-page scroll one RTT later).