Reduce early numbers of two-headed ogres (minmay)
For most characters before Lair, two-headed ogres are difficult monsters
to kill and dangerous to engage in melee. This monster was substantially
buffed in 4a298068, yet its weighting was unadjusted except for
absorbing weight of hill giants when those were removed.

This commit switches the distribution shape of two-headed ogres to SEMI,
which gives them half weight on D:8-11, and also reduces their weight by
10%. These changes mean that about 1.5 two-headed ogres are seen
on-average before D:12, which is about half as many as generate on those
levels presently.