Add orbs of ice to Zot
This commit adds orbs of ice to Zot, in 50% of games. Which games will
have orbs of ice and which will have orbs of fire is fixed on level
generation. This commit breaks backwards compatibility, because it was
necessary to increment the minor version to allow for saving whether a
game is an orb of fire game or an orb of ice game.

Currently, orbs of ice have the same stats as orbs of fire
(with rF+, rC+++ instead of rF+++, rC+) and the spells bolt of cold and
malmutate. They will get iceblast in a subsequent commit, once I have
investigated iceblast's damage formula.

The reasoning behind this commit is that orbs of fire are the primary
reason that fire is rated as the more important element to resist in a 3
rune game. Hence, by making the element the player will have to face at
the end of the game random, we remove this arbitrary decision for fire
to be more important.