Automatically identify all items after the first rune (PR #47)
Late in the game, the identification minigame becomes tedious and doesn't
meaningfully add to gameplay, so, as was voted for, we make identification
automatic once the player obtains their first rune. For now, this identification
occurs on pickup, for logistical reasons. This may change in the future.

The first rune is chosen because it makes the most sense as a concrete
milestone, with somewhat tenable flavour, and by this point the player has
normally identified a large majority of items. This also allows the
retention of challenges such as the Ascetic III tournament banner.

There is no compensation for Ashenzari here; the identification he provided was
80% convenience and only about 20% useful after the first rune point anyway, and
he still allows identification on sight, which has some different gameplay
implications to identification on pickup (for example, inticing players more to
enter sealed vaults).

This commit also removes troves asking for scrolls of identify, because after
the first rune these are now trivial to satisfy.

This commit closes #47, from bhauth.